docker/backendBuild/Dockerfile: upgrade npm
We need to make sure we're using the latest npm so that it can read
our package-lock.json which uses lockfile version 2. In testing, my
local npm handles the new lockfile format correctly, but when I st-
art the backend container the version of npm there is older and it
downgrades the lockfile.