frontend/tsconfig.json: Target ES2015 (aka ES6)
This was the default in Angular 9 anyways¹. ES2015 was released in
2015 so it is currently six years old and supported by all current
major browsers² other than IE 11 (but who cares!).

Another interesting point is that we currently use ESNext in Type-
Script and transpile to ES5, which polyfills newer language features
to their ES5 equivalents. In Google's tests they showed that writing
equivalent ES5 code is much more verbose and therefore slower than
the original ES2015 version.

¹ https://v9.angular.io/guide/typescript-configuration
² https://kangax.github.io/compat-table/es6
³ https://v8.dev/blog/high-performance-es2015