frontend/angular.json: Ignore misleading budget error
After changing our target from ES5 to ES2015 the ng build started to
fail with the following error:

  ERROR in budgets: Exceeded maximum budget for initial-es2015. Budget 5 MB was not met by 6.35 MB with a total of 11.4 MB.
  ERROR in budgets: Exceeded maximum budget for initial-es5. Budget 5 MB was not met by 8.45 MB with a total of 13.5 MB.

There is no doubt that our bundle is large, but in fact it has not
increased as a side effect of the change to ES2015. Here is the main
bundle before and after targeting ES2015:

  $ ls -lh main*.js
  -rw-r--r-- 1 aorth aorth 11M Aug 23 13:10 main.3fcc329c49fba20e2057.js
  -rw-r--r-- 1 aorth aorth 11M Aug 23 13:02 main-es2015.89f7c6f394342cf3dadd.js

The difference is that now Angular also compiles an "initial" bundle
for supporting older browsers, and there is a budget for those files
in angular.json that we previously had not been using.

So the solution here is to delete the budget. Long term we should
see how to reduce the size of that bundle.