frontend: update to Angular 9.1
Minor update to Angular using ng update:

  $ npm install # to make sure packages are present
  $ ng update @angular/cli@^9.1 --allow-dirty
  $ ng update @angular/core@^9.1 --allow-dirty
  $ ng update @angular/cli --from 9 --to 9.1 --allow-dirty --migrate-only
  $ ng update @angular/core --from 9 --to 9.1 --allow-dirty --migrate-only

In this case we didn't have any migrations that needed to be appl-
ied, but I leave it here for documentation.

The Angular 9.1 release is a minor update to 9.0.x and includes a
handful of bugfixes and performance improvements (for example to
ngcc).

See: https://github.com/angular/angular/blob/master/CHANGELOG.md#910-2020-03-25
See: https://blog.angular.io/version-9-1-of-angular-now-available-typescript-3-8-faster-builds-and-more-eb292f989428