backend/tsconfig.json:
Target ES2020 for backend because we're using Node.js 14. We might
be able to use ES2020 modules instead of CommonJS, but that would
need more testing (for example checking the imports).

For this I've tested building, running, and harvesting. Everything
working as expected.

See: https://stackoverflow.com/questions/61305578/what-typescript-configuration-produces-output-closest-to-node-js-14-capabilities/