frontend/package.json: install querystring
This is a Node.js API, not a browser API! Apparently in the past it
was polyfilled by Webpack, and some npm packages got this via the
node-browser-libs package, but since we've updated dependencies we
no longer have it via transitive dependencies and need to install
it explicitly.

According to the Node.js docs, code targetting browsers should be
using URLSearchParams. Note to self...

See: https://nodejs.org/api/querystring.html